Fix recorders list output when pagination is used.
Relying on totalNumberResults in pagination leads to errors since it is
always equal to the total number of recorders irrespective to pagination.
For instance, single device on the page (e.g. on last page when pagination
is in use) will cause error, since it is will be treated as an array.